bootloader_avr_w5500: add support for atmega2560 based boards
Added support for atmega2560 based boards, like the Arduino MEGA
and other compatible devices. This required some changes to account
for the fact that the atmega2560 has 256 KiB of flash and thus
requires some special handling of the EIND register, etc.

+++ bootloader_avr_w5500/src/flash.c
@@ -6,7 +6,7 @@
 
 #include "flash.h"
 
-void flash_page_program(uint16_t addr, uint8_t *page_buf)
+void flash_page_program(uint32_t addr, uint8_t *page_buf)
 {
 	uint16_t i;
 	uint16_t le_word;
